@@ -12,48 +12,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "2.0.0",
-    "@angular/compiler": "2.0.0",
-    "@angular/core": "2.0.0",
-    "@angular/forms": "2.0.0",
-    "@angular/http": "2.0.0",
-    "@angular/platform-browser": "2.0.0",
-    "@angular/platform-browser-dynamic": "2.0.0",
-    "@angular/router": "3.0.0",
-    "@angular2-material/button": "^2.0.0-alpha.8-1",
-    "@angular2-material/card": "^2.0.0-alpha.8-1",
-    "@angular2-material/core": "^2.0.0-alpha.8-1",
-    "@angular2-material/icon": "^2.0.0-alpha.8-1",
-    "@angular2-material/list": "^2.0.0-alpha.8-1",
-    "@angular2-material/sidenav": "^2.0.0-alpha.8-1",
-    "@angular2-material/toolbar": "^2.0.0-alpha.8-1",
-    "@types/firebase": "^2.4.29",
-    "@types/jquery": "^2.0.32",
-    "angularfire2": "^2.0.0-beta.4",
-    "core-js": "^2.4.0",
-    "firebase": "^3.3.2",
-    "gsap": "^1.19.0",
+    "@angular/animations": "^6.0.6",
+    "@angular/common": "6.0.6",
+    "@angular/compiler": "6.0.6",
+    "@angular/core": "6.0.6",
+    "@angular/forms": "6.0.6",
+    "@angular/http": "6.0.6",
+    "@angular/material": "^6.3.0",
+    "@angular/platform-browser": "6.0.6",
+    "@angular/platform-browser-dynamic": "6.0.6",
+    "@angular/router": "6.0.6",
+    "@angular2-material/button": "^2.0.0-alpha.8-3",
+    "@angular2-material/card": "^2.0.0-alpha.8-3",
+    "@angular2-material/core": "^2.0.0-alpha.8-3",
+    "@angular2-material/icon": "^2.0.0-alpha.8-3",
+    "@angular2-material/list": "^2.0.0-alpha.8-3",
+    "@angular2-material/sidenav": "^2.0.0-alpha.8-3",
+    "@angular2-material/toolbar": "^2.0.0-alpha.8-3",
+    "@types/firebase": "^3.2.1",
+    "@types/jquery": "^3.3.4",
+    "angularfire2": "^5.0.0-rc.11",
+    "core-js": "^2.5.7",
+    "firebase": "^5.0.4",
+    "gsap": "^2.0.1",
     "hammerjs": "^2.0.8",
-    "jquery": "^3.1.0",
-    "rxjs": "5.0.0-beta.12",
-    "ts-helpers": "^1.1.1",
-    "zone.js": "0.6.21"
+    "jquery": "^3.3.1",
+    "rxjs": "6.2.1",
+    "ts-helpers": "^1.1.2",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@types/core-js": "^0.9.34",
-    "@types/hammerjs": "^2.0.32",
-    "@types/jasmine": "^2.2.30",
-    "angular-cli": "1.0.0-beta.15",
-    "codelyzer": "~0.0.26",
-    "jasmine-core": "2.4.1",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "0.13.22",
-    "karma-chrome-launcher": "0.2.3",
-    "karma-jasmine": "0.3.8",
-    "karma-remap-istanbul": "^0.2.1",
-    "protractor": "4.0.3",
-    "ts-node": "1.2.1",
-    "tslint": "3.13.0",
-    "typescript": "^2.0.0"
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.6",
+    "@types/core-js": "^2.5.0",
+    "@types/hammerjs": "^2.0.35",
+    "@types/jasmine": "^2.8.8",
+    "codelyzer": "~4.3.0",
+    "jasmine-core": "3.1.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "2.0.4",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-jasmine": "1.1.2",
+    "karma-remap-istanbul": "^0.6.0",
+    "protractor": "5.3.2",
+    "ts-node": "6.1.2",
+    "tslint": "5.10.0",
+    "typescript": "2.7.2"
   }
 }
