@@ -1,10 +1,11 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/observable/merge';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/scan';
-import 'rxjs/add/operator/startWith';
+import { fromEvent, merge } from 'rxjs';
+import { map, scan, startWith } from 'rxjs/internal/operators';
+
+interface Coordinate {
+  x: number,
+  y: number
+}
 
 @Component({
   selector: 'app-merging-streams',
@@ -25,15 +26,17 @@ export class MergingStreamsComponent implements OnInit {
   position: any;
 
   ngOnInit() {
-    const left$ = Observable.fromEvent(this.getNativeElement(this.left), 'click')
-      .map(event => -10);
+    const left$ = fromEvent(this.getNativeElement(this.left), 'click')
+      .pipe(map(event => -10));
 
-    const right$ = Observable.fromEvent(this.getNativeElement(this.right), 'click')
-      .map(event => 10);
+    const right$ = fromEvent(this.getNativeElement(this.right), 'click')
+      .pipe(map(event => 10));
 
-    Observable.merge(left$, right$)
-      .startWith({x: 200, y: 200})
-      .scan((acc, curr) => Object.assign({}, acc, {x: acc.x + curr}))
+    merge(left$, right$)
+      .pipe(
+        startWith({x: 200, y: 200}),
+        scan((acc: Coordinate, curr: number) => Object.assign({}, acc, {x: acc.x + curr}))
+      )
       .subscribe(position => this.position = position);
   }
 
