@@ -1,10 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/observable/merge';
-import 'rxjs/add/operator/mapTo';
-import 'rxjs/add/operator/scan';
-import 'rxjs/add/operator/startWith';
+import { fromEvent, merge, Observable } from 'rxjs';
+import { filter, mapTo, scan, startWith } from 'rxjs/internal/operators';
 
 @Component({
   selector: 'app-map-to-functions',
@@ -21,25 +17,35 @@ export class MapToFunctionsComponent implements OnInit {
   position: any;
 
   ngOnInit() {
-    const leftArrow$ = Observable.fromEvent(document, 'keydown')
-      .filter((event: KeyboardEvent) => event.key === 'ArrowLeft')
-      .mapTo(position => this.decrement(position, 'x', 10));
+    const leftArrow$ = fromEvent(document, 'keydown')
+      .pipe(
+        filter((event: KeyboardEvent) => event.key === 'ArrowLeft'),
+        mapTo(position => this.decrement(position, 'x', 10))
+      )
 
-    const rightArrow$ = Observable.fromEvent(document, 'keydown')
-      .filter((event: KeyboardEvent) => event.key === 'ArrowRight')
-      .mapTo(position => this.increment(position, 'x', 10));
+    const rightArrow$ = fromEvent(document, 'keydown')
+      .pipe(
+        filter((event: KeyboardEvent) => event.key === 'ArrowRight'),
+        mapTo(position => this.increment(position, 'x', 10))
+      )
 
-    const upArrow$ = Observable.fromEvent(document, 'keydown')
-      .filter((event: KeyboardEvent) => event.key === 'ArrowUp')
-      .mapTo(position => this.decrement(position, 'y', 10));
+    const upArrow$ = fromEvent(document, 'keydown')
+      .pipe(
+        filter((event: KeyboardEvent) => event.key === 'ArrowUp'),
+        mapTo(position => this.decrement(position, 'y', 10))
+      )
 
-    const downArrow$ = Observable.fromEvent(document, 'keydown')
-      .filter((event: KeyboardEvent) => event.key === 'ArrowDown')
-      .mapTo(position => this.increment(position, 'y', 10));
+    const downArrow$ = fromEvent(document, 'keydown')
+      .pipe(
+        filter((event: KeyboardEvent) => event.key === 'ArrowDown'),
+        mapTo(position => this.increment(position, 'y', 10))
+      )
 
-    Observable.merge(leftArrow$, rightArrow$, upArrow$, downArrow$)
-      .startWith({x: 100, y: 100})
-      .scan((acc, curr) => curr(acc))
+    merge(leftArrow$, rightArrow$, upArrow$, downArrow$)
+      .pipe(
+        startWith({x: 100, y: 100}),
+        scan((acc, curr: Function) => curr(acc))
+      )
       .subscribe(position => this.position = position);
   }
 
