@@ -1,8 +1,8 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/operator/map';
+import { fromEvent } from 'rxjs';
+
 import 'gsap';
+import { map } from 'rxjs/internal/operators';
 
 @Component({
   selector: 'app-simple-animation',
@@ -16,14 +16,16 @@ export class SimpleAnimationComponent implements OnInit {
     const BALL_OFFSET = 50;
     const CURSOR_OFFSET = 20;
 
-    Observable.fromEvent(document, 'click')
-      .map((event: any) => {
-        const offset = $(event.target).offset();
-        return {
-          x: event.clientX - offset.left - BALL_OFFSET - CURSOR_OFFSET,
-          y: event.pageY - offset.top - BALL_OFFSET - CURSOR_OFFSET
-        };
-      })
+    fromEvent(document, 'click')
+      .pipe(
+        map((event: any) => {
+          const offset = $(event.target).offset();
+          return {
+            x: event.clientX - offset.left - BALL_OFFSET - CURSOR_OFFSET,
+            y: event.pageY - offset.top - BALL_OFFSET - CURSOR_OFFSET
+          };
+        })
+      )
       .subscribe(props => TweenMax.to(this.ball.nativeElement, 1, props))
   }
 }
