@@ -1,10 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { AngularFire } from 'angularfire2';
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/startWith';
+import { fromEvent } from 'rxjs';
+import { scan, startWith } from 'rxjs/internal/operators';
+import { AngularFireDatabase } from 'angularfire2/database';
 
 interface Ticker {
   ticker: number
@@ -19,15 +16,17 @@ interface Ticker {
 export class CounterMasterComponent implements OnInit {
   @ViewChild('btn') btn;
 
-  constructor(private af: AngularFire) {}
+  constructor(private db: AngularFireDatabase) {}
 
   ngOnInit() {
-    const remote$ = this.af.database.object('clicker/');
+    const remoteRef = this.db.object('clicker/');
 
-    Observable.fromEvent(this.getNativeElement(this.btn), 'click')
-      .startWith({ticker: 0})
-      .scan((acc: Ticker, curr) => { return { ticker: acc.ticker + 1 }; })
-      .subscribe(event => remote$.update(event));
+    fromEvent(this.getNativeElement(this.btn), 'click')
+      .pipe(
+        startWith({ticker: 0}),
+        scan((acc: Ticker, curr) => { return { ticker: acc.ticker + 1 }; })
+      )
+      .subscribe(event => remoteRef.update(event));
   }
 
   getNativeElement(element) {
