@@ -1,11 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { AngularFire } from 'angularfire2';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/observable/merge';
-import 'rxjs/add/operator/mapTo';
-import 'rxjs/add/operator/scan';
-import 'rxjs/add/operator/startWith';
+import { fromEvent } from 'rxjs';
+import { map, startWith, switchMap, takeUntil, tap } from 'rxjs/internal/operators';
+import { AngularFireDatabase } from 'angularfire2/database';
 
 @Component({
   selector: 'app-location-master',
@@ -20,35 +16,40 @@ export class LocationMasterComponent implements OnInit {
   @ViewChild('pin') pin;
   position: any;
 
-  constructor(private af: AngularFire) {}
+  constructor(private db: AngularFireDatabase) {}
 
   ngOnInit() {
-    const remote$ = this.af.database.object('location/'),
+    const remoteRef = this.db.object('location/'),
+      remote$ = remoteRef.valueChanges(),
       PIN_OFFSET_X = 26,
       PIN_OFFSET_Y = 16;
 
-    const move$ = Observable.fromEvent(document, 'mousemove')
-      .map((event: MouseEvent) => {
-        const offset = $(event.target).offset();
-        return {
-          x: event.clientX - offset.left - PIN_OFFSET_X,
-          y: event.clientY - offset.top - PIN_OFFSET_Y
-        };
-      });
+    const move$ = fromEvent(document, 'mousemove')
+      .pipe(
+        map((event: MouseEvent) => {
+          const offset = $(event.target).offset();
+          return {
+            x: event.clientX - offset.left - PIN_OFFSET_X,
+            y: event.clientY - offset.top - PIN_OFFSET_Y
+          };
+        })
+      );
 
-    const down$ = Observable.fromEvent(this.pin.nativeElement, 'mousedown')
-      .do(event => this.pin.nativeElement.style.pointerEvents = 'none');
+    const down$ = fromEvent(this.pin.nativeElement, 'mousedown')
+      .pipe(tap(event => this.pin.nativeElement.style.pointerEvents = 'none'));
 
-    const up$ = Observable.fromEvent(document, 'mouseup')
-      .do(event => this.pin.nativeElement.style.pointerEvents = 'all');
+    const up$ = fromEvent(document, 'mouseup')
+      .pipe(tap(event => this.pin.nativeElement.style.pointerEvents = 'all'));
 
     down$
-      .switchMap(event => move$.takeUntil(up$))
-      .startWith({ x: 100, y: 100})
-      .subscribe(event => remote$.update(event));
+      .pipe(
+        switchMap(event => move$.pipe(takeUntil(up$))),
+        startWith({ x: 100, y: 100})
+      )
+      .subscribe(event => remoteRef.update(event));
 
     remote$
-      .startWith({x: 100, y: 100})
+      .pipe(startWith({x: 100, y: 100}))
       .subscribe(result => this.position = result);
   }
 }
