@@ -1,9 +1,7 @@
 import { Component, OnInit, ElementRef } from '@angular/core';
-import { AngularFire } from 'angularfire2';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/pairwise';
+import { fromEvent } from 'rxjs';
+import { map, pairwise, startWith } from 'rxjs/internal/operators';
+import { AngularFireDatabase } from 'angularfire2/database';
 
 declare var jQuery:any;
 
@@ -18,28 +16,31 @@ declare var jQuery:any;
 export class MapMasterComponent implements OnInit {
   lines: any[] = [];
 
-  constructor(private af: AngularFire) {}
+  constructor(private db: AngularFireDatabase) {}
 
   ngOnInit() {
-    const remote$ = this.af.database.object('map/');
+    const remoteRef = this.db.object('map/');
+    const remote$ = remoteRef.valueChanges();
     const emptyLine: any = { x1: 0, y1: 0, x2: 0, y2: 0 };
 
-    Observable.fromEvent(document, 'click')
-      .map((event: MouseEvent) => {
-        const offset = $(event.target).offset();
-        return {
-          x: event.clientX - offset.left,
-          y: event.clientY - offset.top
-        };
-      })
-      .pairwise()
-      .map(positions => {
-        const p1 = positions[0];
-        const p2 = positions[1];
-        return { x1: p1.x, y1: p1.y, x2: p2.x, y2: p2.y };
-      })
-      .startWith(emptyLine)
-      .subscribe(line => remote$.update(line));
+    fromEvent(document, 'click')
+      .pipe(
+        map((event: MouseEvent) => {
+          const offset = $(event.target).offset();
+          return {
+            x: event.clientX - offset.left,
+            y: event.clientY - offset.top
+          };
+        }),
+        pairwise(),
+        map(positions => {
+          const p1 = positions[0];
+          const p2 = positions[1];
+          return { x1: p1.x, y1: p1.y, x2: p2.x, y2: p2.y };
+        }),
+        startWith(emptyLine)
+      )
+      .subscribe(line => remoteRef.update(line));
 
     remote$
       .subscribe(line => this.lines = [...this.lines, line]);
