@@ -1,9 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/interval';
-import 'rxjs/add/operator/timeInterval';
-import 'rxjs/add/operator/take';
-import 'rxjs/add/operator/repeat';
+import { interval } from 'rxjs';
+import { repeat, startWith, take } from 'rxjs/internal/operators';
 
 @Component({
   selector: 'app-game',
@@ -38,11 +35,12 @@ import 'rxjs/add/operator/repeat';
 export class GameComponent implements OnInit{
   backgroundPosition: number = 0;
   ngOnInit() {
-    Observable
-      .interval(10)
-      .startWith(1100)
-      .take(1178)
-      .repeat()
+    interval(10)
+      .pipe(
+        startWith(1100),
+        take(1178),
+        repeat()
+      )
       .subscribe(count => this.backgroundPosition = count);
   }
 }
