@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AngularFire } from 'angularfire2';
+import { AngularFireDatabase } from 'angularfire2/database';
 
 @Component({
   selector: 'app-map-client',
@@ -14,10 +14,10 @@ import { AngularFire } from 'angularfire2';
 export class MapClientComponent implements OnInit {
   lines: any[] = [];
 
-  constructor(private af: AngularFire) {}
+  constructor(private db: AngularFireDatabase) {}
 
   ngOnInit() {
-    const remote$ = this.af.database.object('map/');
+    const remote$ = this.db.object('map/').valueChanges();
 
     remote$
       .subscribe(line => this.lines = [...this.lines, line]);
