@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AngularFire } from 'angularfire2';
+import { AngularFireDatabase } from 'angularfire2/database';
 
 @Component({
   selector: 'app-annotate-client',
@@ -13,10 +13,10 @@ import { AngularFire } from 'angularfire2';
 export class AnnotateClientComponent implements OnInit {
   lines: any[] = [];
 
-  constructor(private af: AngularFire) {}
+  constructor(private db: AngularFireDatabase) {}
 
   ngOnInit() {
-    const remote$ = this.af.database.object('annotate/');
+    const remote$ = this.db.object('annotate/').valueChanges();
 
     remote$
       .subscribe(line => this.lines = [...this.lines, line]);
